Move tools to project, update dependencies
