+++ examples/function_call_assistant/simple_function_call_assistant_duckduckgo_example.py
@@ -2,11 +2,11 @@ import os
 import uuid
 
 from simple_function_call_assistant import SimpleFunctionCallAssistant
-from tools.duckduckgo_tool import DuckDuckGoTool
 
 from grafi.common.containers.container import container
 from grafi.common.models.execution_context import ExecutionContext
 from grafi.common.models.message import Message
+from grafi.tools.functions.impl.duckduckgo_tool import DuckDuckGoTool
 
 
 event_store = container.event_store
